Revert "Task48 develop get method to see the detail of a purchase (#61)"
This reverts commit 9d3180818393fee971010ae74bf6f9c820200989.
