+++ src/AppForSEII2526.API/DTOs/ItemDTOs/ItemForCreateDTO.cs
@@ -7,17 +7,6 @@ namespace AppForSEII2526.API.DTOs.ItemDTOs
         {
         }
 
-        public ItemForCreateDTO(int paymentMethodId, string street, string city, string country, string description, IList<ItemForPurchaseDTO> purchaseItems, decimal totalPrice)
-        {
-            PaymentMethodId = paymentMethodId;
-            Street = street;
-            City = city;
-            Country = country;
-            Description = description;
-            PurchaseItems = purchaseItems;
-            TotalPrice = totalPrice;
-        }
-
         public ItemForCreateDTO(string customerUserName, int paymentMethodId, string street, string city, string country, string description, IList<ItemForPurchaseDTO> purchaseItems, decimal totalPrice)
         {
             CustomerUserName = customerUserName;
