+++ src/AppForSEII2526.API/DTOs/ItemDTOs/ItemForPurchaseDTO.cs
@@ -2,6 +2,9 @@
 {
     public class ItemForPurchaseDTO
     {
+        public ItemForPurchaseDTO()
+        {
+        }
         public ItemForPurchaseDTO(int id, string name, string brand, string description, decimal price, int quantityavailbableforpurchase)
         {
             Id = id;
@@ -12,6 +15,7 @@
             QuantityAvailableForPurchase = quantityavailbableforpurchase;
 
         }
+
 
         public int Id { get; set; }
         [Required]
@@ -26,8 +30,6 @@
         public decimal Price { get; set; }
 
         public int QuantityAvailableForPurchase { get; set; }
-        public int QuantityToBuy { get; set; }
-
 
         public override bool Equals(object? obj)
         {
